Correct program to correctly choose a winner
After a lot of different attempts, I was finally able to get
the code to do what it needed to do.

The main issue I was having was not getting the correct result from
getComputerChoice() which would tell me playerSelection won when it
should have lost.

This was because I was basing it off the console.log after the
getComputerChoice() function and then again at playRound, which would
sometimes give me a different result.